requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
apscheduler
jinja2
python-multipart
requests
beautifulsoup4
lxml
pytz
python-dotenv
